Nozzle tile: selective whitelist of KPI/KCSI metrics.
Stackdriver Monitoring enforces a limit of 500 custom metrics per project.
Many of the metrics the nozzle automatically derives from the firehose are
not useful for understanding the performance of a PCF instance, and thus
wastefully consume a scarce resource. Pivotal's documentation describes
the set of firehose metrics that *are* useful, so the tile whitelists
these by default.

Note that installing the BOSH release directly into the director will
not set up this whitelisting.